% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGP_Mom.R
\name{FitGP_Mom}
\alias{FitGP_Mom}
\title{FitGP_Mom}
\usage{
Value <- FitGP_Mom(X, p, N= 0, r11= 1, fixedpar= NULL, l0= NULL, sigma= Inf, XId= '')
}
\arguments{
\item{X}{data sample (double(n))}

\item{p}{probabilities of exceedance of the quantiles to be estimated (double(np))}

\item{N}{(optional) (effective) sample size, in case X is not complete but contains only (peak) values above some threshold (integer(1))}

\item{r11}{(optional) factor to increase estimator variance by, to account for serial dependence (default: 1) (double(1) or list, see Details)}

\item{fixedpar}{(optional): fixed model parameters not to be estimated, and their standard errors (double(1) or list, see Details)}

\item{l0}{(optional) value of l (no. of order stats used) in case it is imposed (integer(0))}

\item{sigma}{(optional) determines the ratio of k to l, see below (double(1))}

\item{XId}{(optional) data identifier to store with output for traceability (character)}
}
\value{
A list, with members: 
  \item{l}{no. of order statistics used for scale and quantile estimation}    
  \item{k}{no. of order statistics used for tail index estimation} 
  \item{tailindex}{estimates or imposed value of GP tail index} 
  \item{tailindexStd}{standard deviations of tail index estimates}
  \item{logdisp}{estimates or imposed value of log of dispersion coeff.}  
  \item{logdispStd}{standard deviations of log of dispersion coeff. estimates}
  \item{scale}{estimates of GP scale parameter}
  \item{locationStd}{standard deviation of order statistic}
  \item{lambda}{ratio of logarithms of probabilities of exceedance of quantile and threshold}  
  \item{p}{probabilities of exceedance of quantiles to be estimated} 
  \item{quantile}{quantile estimates}
  \item{quantileStd}{standard deviations of quantile estimates}
  \item{orderstats}{data X sorted (decreasing)}
  \item{df}{= "GP": fitted distribution function tail (Generalised Pareto)}
  \item{estimator}{= "maximum likelihood": see "method" below}
}
\description{
Fit GP tail using moment method (see reference)
}
\details{
Pre-determined model parameters are to be supplied in the list fixedpar (see above):
 \itemize{
  \item{$gamma0: (optional) value of tailindex in case it is imposed (double(1))}
  \item{$gamma0Std: (optional) its standard deviation (double(1))}
  \item{$logdisp0: (optional) value of log of dispersion coeff. in case it is imposed (dispersion coeff. is the raio of scale par. to location par.) (double(1))}
  \item{$logdisp0Std: (optional) its standard deviation (double(1))}        
  }
  
  The serial dependence coefficient r11 can be a positive number, or a list 
  produced by R11.R. 
  
  In case a quantile is to be estimated for a \emph{frequency}, say f, and 
  \enumerate{
  \item{if X contains all values (possibly above some threshold), then with
  EI an estimate of the Extremal Index from EI.R, set
  p = f*d/EI and N = T/d, with T the length of the observation period and d the time step. 
        Note that f and d are defined with reference to the same unit of time!! In this case,
        r11 needs to be estimated.
       }
  \item{if X contains only the n (approximately Poisson) peak values above some threshold 
        (in a PoT analysis),  it is recommended to set r11= 1 and take p = f*d/EI and 
        N = T/d*EI; in this case (for GP), EI can be any value; e.g. take p= fT/n and N= n.
       } 
}
}
\references{
De Haan, L. and A. Ferreira (2006), Extreme Value Theory - An Introduction. Springer.
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
