% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGW_iHill.R
\name{FitGW_iHill}
\alias{FitGW_iHill}
\title{Fit GW Hill}
\usage{
FitGW_iHill(X, p, N, EI, theta0, logdisp0, l0, XId)
}
\arguments{
\item{X}{double(n)   data sample}

\item{p}{probabilities of exceedance of quantiles to be estimated}
}
\description{
Fit a Generalised Weibull (GW) tail to the sample X and
}
\author{
Cees de Valk, KNMI
}
