% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGW_iHillpos.R
\name{FitGW_iHillpos}
\alias{FitGW_iHillpos}
\title{FitGW_iHillpos}
\usage{
Value <- FitGW_iHillpos(X, p, N= 0, r11= 1, fixedpar= NULL, l0= NULL, sigma= 1, indexsign= 0, XId= '')
}
\arguments{
\item{X}{data sample (double(n))}

\item{p}{probabilities of exceedance of the quantiles to be estimated (double(np))}

\item{N}{(optional) (effective) sample size, in case X is not complete but contains only (peak) values above some threshold (integer(1))}

\item{r11}{(optional) factor to increase estimator variance by, to account for serial dependence (default: 1) (double(1) or list, see Details)}

\item{fixedpar}{(optional): fixed model parameters not to be estimated, and their standard errors (list; see Details)}

\item{l0}{(optional) value of l (no. of order stats used) in case it is imposed (integer(0))}

\item{sigma}{(optional) determines the ratio of k to l (double(1))}

\item{indexsign}{(optional) direction of sorting of dispersion coefficient. If 0, Weibull fit returned}

\item{XId}{(optional) data identifier to store with output for traceability (character)}
}
\value{
A list, with members: 
  \item{l}{no. of order statistics used for scale and quantile estimation}    
  \item{k}{no. of order statistics used for tail index estimation} 
  \item{sigma}{algorithm parameter (see ref. eq. (30))}
  \item{tailindex}{estimates or imposed value of GW tail index} 
  \item{tailindexStd}{standard deviations of tail index estimates}
  \item{logdisp}{estimates or imposed value of log of dispersion coeff.}  
  \item{logdispStd}{standard deviations of log of dispersion coeff. estimates}
  \item{scale}{estimates of GW scale parameter}
  \item{locationStd}{standard deviation of order statistic}
  \item{lambda}{ratio of logarithms of probabilities of exceedance of quantile and threshold}  
  \item{p}{probabilities of exceedance of quantiles to be estimated} 
  \item{quantile}{quantile estimates}
  \item{quantileStd}{standard deviations of quantile estimates}
  \item{tailindexraw}{estimates of GW tail index over all possible thresholds} 
  \item{tailindexrawStd}{standard deviation of tailindexraw}
  \item{kraw}{no. of order statistics used for estimation of tailindexraw} 
  \item{orderstats}{data X sorted (decreasing)}
  \item{df}{= "GW": fitted distribution function tail (Generalised Weibull}
  \item{estimator}{= "iteratedHill": see "method" below}
}
\description{
Fit a Generalised Weibull (GW) upper tail to the sample X and estimate quantiles; X 
must be positive with nonzero probability.
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
