% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTail_PoT.R
\name{FitTail_PoT}
\alias{FitTail_PoT}
\title{FitTail_PoT}
\usage{
Value <- FitTail_PoT(X, freq= NULL, df= "GW", method= "FitGW_iHilli", options= NULL, metadata= NULL)
}
\arguments{
\item{X}{data sample (double(n))}

\item{freq}{frequencies of exceedance of the quantiles to be estimated (double(nf))}

\item{df}{distribution function to be fitted to the tail: "GP", "GW", "logGW", "Wbl", or "Exp"}

\item{method}{(optional) name of R-script to estimate the tail (character)}

\item{options}{(optional) parameters controlling the estimation (list; see Details)}

\item{metadata}{(optional) information about the variable and the time-series (list; see Details)}
}
\value{
A list, with members: 
  \item{l}{no. of order statistics used for scale and quantile estimation}    
  \item{k}{no. of order statistics used for tail index estimation} 
  \item{sigma}{algorithm parameter (see ref. eq. (30))}
  \item{tailindex}{estimates or imposed value of GW tail index} 
  \item{tailindexStd}{standard deviations of tail index estimates}
  \item{logdisp}{estimates or imposed value of log of dispersion coeff.}  
  \item{logdispStd}{standard deviations of log of dispersion coeff. estimates}
  \item{scale}{estimates of GW scale parameter}
  \item{locationStd}{standard deviation of order statistic}
  \item{lambda}{ratio of logarithms of probabilities of exceedance of quantile and threshold}  
  \item{p}{probabilities (frwctions of time) of exceedance of quantiles to be estimated} 
  \item{p0}{probability of exceedance of the lower bound of the variable} 
  \item{freq}{frequencies of exceedance of quantiles to be estimated}    
  \item{quantile}{quantile estimates}
  \item{quantileStd}{standard deviations of quantile estimates}
  \item{tailindexraw}{raw estimates of GW tail index over all possible thresholds (method: FitGW_iHill.R)} 
  \item{tailindexrawStd}{standard deviation of tailindexraw}
  \item{kraw}{no. of order statistics used for estimation of tailindexraw} 
  \item{orderstats}{data X sorted (decreasing)}
  \item{df}{see above}
  \item{method}{see above}

 options may contain the following fields:
 \itemize{
  \item{$pthreshold: fraction of time that value exceeds threshold (double(1))}
  \item{$maxpthreshold: upper bound on pthreshold (in case pthreshold is estimated)}
  \item{$minpthreshold: lower bound on pthreshold (in case pthreshold is estimated)}
  \item{$indexselect: if TRUE, threshold is selected based on tail index estimates (logical, default= FALSE)} 
  \item{$kmin: no. of order statistics skipped in determining threshold (integer(1)), default= 20)} 
  \item{$sigma: determines the ratio of k to l ( (no. of order stats used for estimation of tail index and quantile) (double(1)}
  \item{$fixedpar: fixed model parameters not to be estimated, and their standard errors (list; see below)}
  \item{$plotparams: plotparameters (list) with members: $makeplot (default= TRUE), $pconf (coverage probability of confidence interval), $xlim (plot limits for quantile estimates), $freqlim (plot limits for frequencies), $plim (plot limits for fractions of time)}
 }
}
\description{
Fit a tail to a PoT (peaks-over-threshold) sample from a time series, and estimate upper tail quantiles
}
\author{
Cees de Valk \email{cees.de.valk@knmi.nl}
}
