% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTailFromAllData.R
\name{FitTailFromAllData}
\alias{FitTailFromAllData}
\title{FitTailFromAllData}
\usage{
Value <- FitTailFromAllData(X, freq= NULL, df= "GW", method= "FitGW_iHilli", options= NULL, metadata= NULL)
}
\arguments{
\item{X}{data sample (double(n), optionally with a second column of values of a covariate (double(n, 2)))}

\item{freq}{frequencies of exceedance of the quantiles to be estimated (double(nf))}

\item{df}{distribution function to be fitted to the tail: "GP", "GW", "logGW", "Wbl", or "Exp"}

\item{method}{(optional) name of R-script to estimate the tail (character)}

\item{options}{(optional) parameters controlling the estimation (list; see Details)}

\item{metadata}{(optional) information about the variable and the time-series (list; see Details)}
}
\value{
A list, with members: 
  \item{l}{no. of order statistics used for scale and quantile estimation}    
  \item{k}{no. of order statistics used for tail index estimation} 
  \item{sigma}{algorithm parameter (see ref. eq. (30))}
  \item{tailindex}{estimates or imposed value of GW tail index} 
  \item{tailindexStd}{standard deviations of tail index estimates}
  \item{logdisp}{estimates or imposed value of log of dispersion coeff.}  
  \item{logdispStd}{standard deviations of log of dispersion coeff. estimates}
  \item{scale}{estimates of GW scale parameter}
  \item{locationStd}{standard deviation of order statistic}
  \item{lambda}{ratio of logarithms of probabilities of exceedance of quantile and threshold}  
  \item{p}{probabilities (fractions of time) of exceedance of quantiles to be estimated} 
  \item{pbin}{probability (fraction of time) of exceedance of the lower bound of the variable} 
  \item{freq}{frequencies of exceedance of quantiles to be estimated}    
  \item{quantile}{quantile estimates}
  \item{quantileStd}{standard deviations of quantile estimates}
  \item{tailindexraw}{raw estimates of GW tail index over all possible thresholds (method: FitGW_iHill.R)} 
  \item{tailindexrawStd}{standard deviation of tailindexraw}
  \item{kraw}{no. of order statistics used for estimation of tailindexraw} 
  \item{orderstats}{data X sorted (decreasing)}
  \item{df}{see above}
  \item{method}{see above}
  \item{ }{In addition, several plots are produced (tailindex, quantile, threshold indicator, fitted tail for selected threshold)}
}
\description{
Fit and extrapolate the tail of the frequency distribution of a time-series. Optionally, the distribution of the values coinciding with a covariate in some bin is estimated, for each of a specified set of bins.
}
\details{
metadata may contain the following fields (in addition to your own meta data):
 \itemize{
  \item{caseId: user-chosen identifier for later reference (default: current date/time)}
  \item{$varname: variable name}
  \item{$varunit: physical unit of variable}
  \item{$timeunit: time unit (e.g. year)}
  \item{$timestep: time step in units of timeunit}
  \item{$timelength: length of time covered by time-series, in units of timeunit} 
  \item{$nexcess (for PoT only): no. of data values (as opposed to peak values) exceeding the threshold}
 }  
 
 options may contain the following fields:
 \itemize{
  \item{$covariate: list, which may contain $lbin and $ubin (the lower resp. upper limits of the bins of covariate values on which to restrict the tail estimates)}
  \item{$dither: width of uniform distribution of noise to add to data (double(1))}
  \item{$pthreshold: fraction of good values exceeding threshold (double(1))}
  \item{$maxpthreshold: upper bound on pthreshold (in case pthreshold is estimated)}
  \item{$minpthreshold: lower bound on pthreshold (in case pthreshold is estimated)}
  \item{$indexselect: if TRUE, threshold is selected based on tail index estimates (logical, default= FALSE)} 
  \item{$kmin: no. of order statistics skipped in determining threshold (integer(1)), default= 20)} 
  \item{$sigma: determines the ratio of k to l ( (no. of order stats used for estimation of tail index and quantile) (double(1)}
  \item{$bootstrap: list. If exists/nonempty, precision is assessed by a moving block bootstrap. May contain $blocktime (block length in terms of time) and $nsamples (no. of bootstrap samples) or $blocks (random starting indices of the blocks, an array of size ($nblocks, $nsamples))}
  \item{$fixedpar: fixed model parameters not to be estimated, and their standard errors (list; see below)}
  \item{$plotparams: plotparameters (list) with members: $makeplot (default= TRUE), $pconf (coverage probability of confidence interval), $xlim (plot limits for quantile estimates), $freqlim (plot limits for frequencies), $plim (plot limits for fractions of time)}
 }   

 Pre-determined model parameters are to be supplied in options$fixedpar (see above):
 \itemize{
  \item{$theta0: (optional) value of tailindex in case it is imposed (double(1))}
  \item{$theta0Std: (optional) its standard deviation (double(1))}
  \item{$logdisp0: (optional) value of log of dispersion coeff. in case it is imposed (dispersion coeff. is the raio of scale par. to location par.) (double(1))}
  \item{$logdisp0Std: (optional) its standard deviation (double(1))}        
  }
}
\author{
Cees de Valk \email{cees.de.valk@knmi.nl}
}
