% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitWbl_MLE.R
\name{FitWbl_MLE}
\alias{FitWbl_MLE}
\title{FitWbl_MLE}
\usage{
Value <- FitWbl_MLE(X, p, N= 0, r11= 1, fixedpar= NULL, l0= NULL, sigma= 1, XId= '')
}
\arguments{
\item{X}{data sample (double(n))}

\item{p}{probabilities of exceedance of the quantiles to be estimated (double(np))}

\item{N}{(optional) (effective) sample size, in case X is not complete but contains only (peak) values above some threshold (integer(1))}

\item{r11}{(optional) factor to increase estimator variance by, to account for serial dependence (default: 1) (double(1) or list, see Details)}

\item{fixedpar}{(optional): fixed model parameters not to be estimated, and their standard errors (list; see Details)}

\item{l0}{(optional) value of l (no. of order stats used) in case it is imposed (integer(0))}

\item{sigma}{(optional) fixed algorithm parameter (see de Valk & Cai (2018) eq. (30)) (double(1))}

\item{XId}{(optional) data identifier to store with output for traceability (character)}
}
\value{
A list, with members: 
  \item{l}{no. of order statistics used for scale and quantile estimation}    
  \item{k}{no. of order statistics used for tail index estimation} 
  \item{sigma}{fixed algorithm parameter (see ref. eq. (30))}
  \item{tailindex}{estimates or imposed value of GW tail index} 
  \item{tailindexStd}{standard deviations of tail index estimates}
  \item{f}{estimates or imposed value of offset/location.}  
  \item{fStd}{standard deviations of offset/location estimates}
  \item{locationStd}{standard deviation of order statistic}
  \item{lambda}{ratio of logarithms of probabilities of exceedance of quantile and threshold}  
  \item{p}{probabilities of exceedance of quantiles to be estimated} 
  \item{quantile}{quantile estimates}
  \item{quantileStd}{standard deviations of quantile estimates}
  \item{orderstats}{data X sorted (decreasing)}
  \item{df}{= "Weibull": fitted tail}
  \item{estimator}{= "MLE": see "method" below}
}
\description{
Fit a Generalised Weibull (GW) upper tail to the sample X and estimate quantiles, using
the ML estimator for tail index and scale
}
\details{
Pre-determined model parameters are to be supplied in the list fixedpar (see above):
 \itemize{
  \item{$theta0: (optional) value of tailindex in case it is imposed (double(1))}
  \item{$theta0Std: (optional) its standard deviation (double(1))}
  \item{$fp0: (optional) value of of offset/location in case it is imposed (dispersion coeff. is the raio of scale par. to location par.) (double(1))}
  \item{$f0Std: (optional) its standard deviation (double(1))}        
  }
  
  The serial dependence coefficient r11 can be a positive number, or a list 
  produced by R11.R. 
  
  In case a quantile is to be estimated for a \emph{frequency}, say f, and 
  \enumerate{
  \item{if X contains all values (possibly above some threshold), then with
  EI an estimate of the Extremal Index from EI.R, set
  p = f*d/EI and N = T/d, with T the length of the observation period and d the time step. 
        Note that f and d are defined with reference to the same unit of time!! In this case,
        r11 needs to be estimated.
            }
  \item{if X contains only the (approximately Poisson) peak values above some threshold 
        (i.e., you want to do a PoT analysis), then set p = f*d/EI, N = (T/d)*EI, r11= 1 
        (note that d/EI is the mean duration of an "event" and EI/d is the 
        mean number of "events" per unit of time estimated by EI.R).
             }
}
}
\references{
?????????????
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
