% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qckGW.R
\name{qckGW}
\alias{qckGW}
\title{qckGW}
\usage{
Value <- qckGW(p, es, l)
}
\arguments{
\item{p}{probability(ies) of exceedance at which to evaluate the quantile(s) (double(np))}

\item{es}{fitted GW model (see e.g. FitGW_iHill.R)}

\item{l}{(optional) threshold rank: no. of order statistics exceeding the threshold for 
the scale parameter (see ref. for notation)}
}
\value{
list containing quantile(s) and optionally, their standard deviations
}
\description{
quantile of the GW (Generalised Weibull) tail and its standard deviation
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
