% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGW.R, R/qGWdisp.R
\name{qGW}
\alias{qGW}
\alias{qGW}
\alias{qGWdisp}
\title{qGW}
\usage{
Value <- QGW(p, p0, q0, scale, theta, q0Std, scaleStd, thetaStd)

Value <- QGW(p, p0, q0, logdisp, theta, q0Std, logdispStd, thetaStd)
}
\arguments{
\item{p}{probability(ies) of exceedance at which to evaluate the quantile(s) (double(np))}

\item{p0}{reference probability for which the quantile q0 is given (double(1))}

\item{q0}{see under p0}

\item{scale}{GW scale parameter at p0 (double(1))}

\item{theta}{GW tail index (double(1))}

\item{q0Std}{(optional) standard deviation of q0 (double(1))}

\item{scaleStd}{(optional) standard deviation of scale (double(1))}

\item{thetaStd}{(optional) standard deviation of theta (double(1))}

\item{logdisp}{logarithm of the dispersion coefficient; the latter is scale/q0 (double(1))}

\item{logdispStd}{(optional) standard deviation of logdisp}
}
\value{
list containing quantile(s) and optionally, their standard deviations

list containing quantile(s) and optionally, their standard deviations
}
\description{
quantile of the GW (Generalised Weibull) tail and (optionally) its standard deviation

As qGW, GW scale parameter scale is replaced by the logarithm of the disperson coefficient
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}

De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}

Cees de Valk \email{ceesfdevalk@gmail.com}
}
