% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGWdisp.R
\name{qGWdisp}
\alias{qGWdisp}
\title{qGWdisp}
\usage{
Value <- QGW(p, p0, q0, logdisp, theta, q0Std, logdispStd, thetaStd)
}
\arguments{
\item{logdisp}{logarithm of the dispersion coefficient scale/q0 (double(1))}

\item{logdispStd}{(optional) standard deviation of logdisp}
}
\value{
list containing quantile(s) and optionally, their standard deviations
}
\description{
As qGW, but the GW scale parameter "scale" is replaced by the logarithm of the disperson coefficient
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
