% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r11.R
\name{r11}
\alias{r11}
\title{r11}
\usage{
Value <- r11(X, l, ngr, makeplot)
}
\arguments{
\item{X}{time series (double(nX))}

\item{l}{(optional, defailt= 50) no of data points above highest level (integer(1))}

\item{ngr}{(optional, default= 10) number of levels (integer(1))}

\item{makeplot}{(optional, default= FALSE) (logical(1))}
}
\value{
A list, with members: 
  \item{k}{no. of data points above level, for each of ngr levels}
  \item{p}{sample fraction above level, for each of ngr levels} 
  \item{r}{corresponding estimate of r(1,1)}
}
\description{
Computes estimates of the extremal dependence coeficient r(1,1) of 
Rootzen (1995)
}
\references{
Rootzen, H. (1995), The tail empirical process for stationary sequences. Technical report,
 Department of Mathematics, Chalmers University, Sweden.
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
