% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectThresholdBT.R
\name{selectThresholdBT}
\alias{selectThresholdBT}
\title{selectThresholdBT}
\usage{
Value <- selectThresholdBT(tailindex, tailindexStd, l, l0)
}
\arguments{
\item{l}{number of order statistics above the threshold for scale
and location estimation from FitGW_iHill.R (double(nl))}

\item{l0}{(optional) lower bound of range of l to check}

\item{tailindex}{GW tail index estimates from FitGW_iHill.R (double(nl))}

\item{tailindexStd}{Standard deviation of tail index from FitGW_iHill.R (double(nl))}
}
\description{
Automatic threshold selection for GW (Generalised Weibull) tail 
             estimation, based on fluctuation statistics of the GW tail index
}
\details{
Works together with FitGW_iHill.M. 
       Based on ref. De Valk and Cai(2018), in particular on eq.(27).
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
