% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectThresholdP0.R
\name{selectThresholdP0}
\alias{selectThresholdP0}
\title{selectThresholdP0}
\usage{
Value <- selectThresholdP0(tailindex, tailindexStd)
}
\arguments{
\item{l}{number of order statistics above the threshold for scale
and location estimation from FitGW_iHill.R (double(nl))}

\item{tailindex}{GW tail index estimates from FitGW_iHill.R (double(nl))}

\item{tailindexStd}{Standard deviation of tail index from FitGW_iHill.R (double(nl))}
}
\value{
list containing the elements
  \item{i}{index in the vectors tailindex and l, representing the selected threshold}    
  \item{P}{p-values of observed fluctuation statistic}    
  \item{bias}{estimate of bias in tail index based on fluctuation statistic}
}
\description{
Automatic threshold selection for GW (Generalised Weibull) tail 
             estimation, based on fluctuation statistics of the GW tail index
}
\details{
Works together with FitGW_iHill.M. 
       Based on ref. De Valk and Cai(2018), in particular on eq.(27).
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
