% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectThresholdP1.R
\name{selectThresholdP1}
\alias{selectThresholdP1}
\title{selectThresholdP1}
\usage{
Value <- selectThresholdP1(tailindex, tailindexStd, k, rthresh)
}
\arguments{
\item{k}{number of order statistics above the threshold for 
estimation of tail index from FitGW_iHill.R (double(nl))}

\item{rthresh}{(optional) ratio of the probability of nonexceedance of fluctuation 
size to its maximum, for threshold to be accepted. Default is 0.5.}

\item{tailindex}{GW tail index estimates from FitGW_iHill.R (double(nl))}

\item{tailindexStd}{Standard deviation of tail index from FitGW_iHill.R (double(nl))}
}
\value{
list containing the elements
  \item{i}{index in the vectors tailindex and l, representing the selected threshold}  
  \item{k}{the elements of k for which P and bias are computed}      
  \item{P}{p-values of observed fluctuation statistic}    
  \item{bias}{estimate of bias in tail index based on fluctuation statistic}
}
\description{
Automatic threshold selection for GW (Generalised Weibull) tail 
             estimation, based on fluctuation statistics of the GW tail index
}
\details{
Works together with FitGW_iHill.R and similar methods.  
       Based on ref. De Valk and Cai(2018), in particular on eq.(27).
       Idea of simple estimate is from Boucheron & Thomas (2015); see also  
       Drees and Kaufmann (1998).
       Phi in this equation (see (22)) can be approximated by a rescaled Brownian
       motion in the large-sample limit; then we use the BM statistics as in
       Lemma 1 of de Valk & Cai (2018).
       Serial dependence is accounted for in a simplistic way implicitly through 
       tailindexStd.
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128.
Boucheron, S., Thomas, M. (2015), Tail index estimation, concentration and adaptivity. 
Electron. J. Stat. 9, 2751–2792.
Drees, H., Kaufmann, E. (1998), Selecting the optimal sample fraction in univariate 
extreme value estimation. Stoch. Process. Appl. 75, 149–172.
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
