% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindexplot.R
\name{tailindexplot}
\alias{tailindexplot}
\title{tailindexplot}
\usage{
tailindexplot(es, params)
}
\arguments{
\item{es}{list containing tail estimates from a single sample}

\item{params}{(optional) list (see below)}
}
\value{
A plot file (.png)
}
\description{
# Plot of tail index estimates with confidence interval
}
\details{
The parameter list params may contain:
 \itemize{
  \item{$pconf: coverage probability of confidence interval (0.9 by default) (double(1))}
  \item{$plim: plot limits for sample fraction (double(2))}  
  \item{$tailindexlim: plot limits for tail index (double(2))}  
  }
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
