% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailquantileplot.R
\name{tailquantileplot}
\alias{tailquantileplot}
\title{tailquantileplot}
\usage{
tailquantileplot(estimates, params)
}
\arguments{
\item{params}{(optional) list (see below)}

\item{estimates}{list containing tail estimates from a single sample}
}
\value{
A plot file (.png)
}
\description{
# Plot of tail quantile estimates for a single frequency with confidence interval
}
\details{
The parameter list params may contain:
 \itemize{
  \item{$pconf: coverage probability of confidence interval (0.9 by default) (double(1))}
  }
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
