% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdplot.R
\name{thresholdplot}
\alias{thresholdplot}
\title{thresholdplot}
\usage{
thresholdplot(params, es)
}
\arguments{
\item{params}{(optional) list (see below)}

\item{es}{list containing tail estimates from a single sample}
}
\value{
A plot file (.png)
}
\description{
# Plot of p-value of a fluctuation size statistic, for threshold selection
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
