% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGW_iHill.R
\name{FitGW_iHill}
\alias{FitGW_iHill}
\title{FitGW_iHill}
\usage{
Value <- FitGW_iHill(X, p, N= 0, r11= 1, fixedpar= NULL, l0= NULL, sigma= 1, XId= '')
}
\arguments{
\item{X}{data sample (double(n))}

\item{p}{probabilities of exceedance of the quantiles to be estimated (double(np))}

\item{N}{(optional) (effective) sample size, in case X is not complete but contains only (peak) values above some threshold (integer(1))}

\item{r11}{(optional) factor to increase estimator variance by, to account for serial dependence (default: 1) (double(1))}

\item{fixedpar}{(optional): fixed model parameters not to be estimated, and their standard errors (list; see Details)}

\item{l0}{(optional) value of l (no. of order stats used) in case it is imposed (integer(0))}

\item{sigma}{(optional) determines the ratio of k to l (double(1))}

\item{XId}{(optional) data identifier to store with output for traceability (character)}
}
\value{
A list, with members: 
  \item{l}{no. of order statistics used for scale and quantile estimation}    
  \item{k}{no. of order statistics used for tail index estimation} 
  \item{sigma}{algorithm parameter (see ref. eq. (30))}
  \item{tailindex}{estimates or imposed value of GW tail index} 
  \item{tailindexStd}{standard deviations of tail index estimates}
  \item{logdisp}{estimates or imposed value of log of dispersion coeff.}  
  \item{logdispStd}{standard deviations of log of dispersion coeff. estimates}
  \item{scale}{estimates of GW scale parameter}
  \item{locationStd}{standard deviation of order statistic}
  \item{lambda}{ratio of logarithms of probabilities of exceedance of quantile and threshold}  
  \item{p}{probabilities of exceedance of quantiles to be estimated} 
  \item{quantile}{quantile estimates}
  \item{quantileStd}{standard deviations of quantile estimates}
  \item{orderstats}{data X sorted (decreasing)}
  \item{df}{= "GW": fitted distribution function tail (Generalised Weibull}
  \item{estimator}{= "iteratedHill": see "method" below}
}
\description{
Fit a Generalised Weibull (GW) upper tail to the sample X and estimate quantiles
}
\details{
Pre-determined model parameters are to be supplied in the list fixedpar (see above):
 \itemize{
  \item{$theta0: (optional) value of tailindex in case it is imposed (double(1))}
  \item{$theta0Std: (optional) its standard deviation (double(1))}
  \item{$logdisp0: (optional) value of log of dispersion coeff. in case it is imposed (dispersion coeff. is the raio of scale par. to location par.) (double(1))}
  \item{$logdisp0Std: (optional) its standard deviation (double(1))}        
  }
  
  In case a quantile is to be estimated for a \emph{frequency}, say f, and 
  \enumerate{
  \item{if X contains all values (possibly above some threshold), then with
  EI an estimate of the Extremal Index, set
  p = f*d/EI and N = T/d, with T the length of the observation period and d the time step. 
        Note that f and d are defined with reference to the same unit of time!! In this case,
        r11 needs to be estimated.
            }
  \item{if X contains only the (approximately Poisson) peak values above some threshold 
        (i.e., you want to do a PoT analysis), then set p = f*d/EI, N = (T/d)*EI, r11= 1 
        (note that d/EI is the mean duration of an "event" and EI/d is the 
        mean number of "events" per unit of time).
             }
}
}
\references{
De Valk, C. and Cai, J.J. (2018), A high quantile estimator based on 
the log-generalized Weibull tail limit. Econometrics and Statistics 6, 107-128, see
\url{https://doi.org/10.1016/j.ecosta.2017.03.001}
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
