% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTail_AllData.R
\name{FitTail_AllData}
\alias{FitTail_AllData}
\title{FitTail_AllData}
\usage{
Value <- FitTail_AllData(X, freq= NULL, df= "GW", method= "FitGW_iHilli", options= NULL, metadata= NULL)
}
\arguments{
\item{X}{data sample (double(n) or (double(n, 2)))}

\item{freq}{frequencies of exceedance of the quantiles to be estimated (double(nf))}

\item{df}{distribution function to be fitted to the tail: "GP", "GW", "logGW", "Wbl", or "Exp"}

\item{method}{(optional) name of R-script to estimate the tail (character)}

\item{options}{(optional) parameters controlling the estimation (list; see Details)}

\item{metadata}{(optional) information about the variable and the time-series (list; see Details)}
}
\description{
Fit a tail to the values of a time-series and estimate upper tail quantiles
}
\author{
Cees de Valk \email{cees.de.valk@knmi.nl}
}
