% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoTselect.R, R/PoTtselect.R
\name{PoTselect}
\alias{PoTselect}
\title{PoTselect}
\usage{
XPoT <- PoTselect(X, p, separation)

XPoT <- PoTselect(X, p, separation)
}
\arguments{
\item{X}{time series (sequence) (double(n))}

\item{p}{fraction of elements of X exceeding the threshold (double(1))}

\item{separation}{minimum distance between selected peaks in the sequence (integer(1))}
}
\value{
list containing the elements
  \item{pot}{selected peaks} 
  \item{ind}{indices in the sequence of the selected peaks in pot}  
  \item{p}{same as input}

list containing the elements
  \item{pot}{selected peaks} 
  \item{ind}{indices in the sequence of the selected peaks in pot}  
  \item{p}{same as input}
}
\description{
Select a Peak over Threshold (PoT) sample from a univariate time series

Select a Peak over Threshold (PoT) sample from a univariate time series
}
\author{
Alexandre Tribut, Cees de Valk \email{ceesfdevalk@gmail.com}

Cees de Valk \email{ceesfdevalk@gmail.com}
}
