% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailplot.R
\name{tailplot}
\alias{tailplot}
\title{tailplot}
\usage{
tailplot(es1, es2, ..., params= params)
}
\arguments{
\item{...}{one or several lists es1, es2, ..., each containing a tail estimate to be plotted}

\item{params}{(optional) list (see below)}
}
\value{
A plot file (.png)
}
\description{
# Plot of tail estimates for one or several samples, with confidence interval
}
\details{
The parameter list params must be explicitly specified with "params= ..." and may contain:
 \itemize{
  \item{$pconf: coverage probability of confidence interval (0.9 by default) (double(1))}
  \item{$xlim: plot limits for quantile estimates (double(2))}
  \item{$freqlim: plot limits for frequency (double(2))}
  }
}
\author{
Cees de Valk \email{ceesfdevalk@gmail.com}
}
